Eliminate code duplication in expression builders, minor cleanup.
