+++ mcfc/cudaexpression.py
@@ -20,8 +20,6 @@
 from expression import *
 from cudaparameters import numElements
 
-import ufl.differentiation
-
 # Variables
 
 threadCount = Variable("THREAD_COUNT")
@@ -44,8 +42,7 @@ def buildElementLoop():
     init = InitialisationOp(var, threadId)
     test = LessThanOp(var, numElements)
     inc = PlusAssignmentOp(var, threadCount)
-    ast = ForLoop(init, test, inc)
-    return ast
+    return ForLoop(init, test, inc)
 
 def buildSubscript(variable, indices):
     """Given a list of indices, return an AST that computes
@@ -75,7 +72,6 @@ def buildMultiArraySubscript(variable, indices):
     subscripted by the indices as if it were a multidimensional
     array."""
 
-    indexList = []
     for i in indices:
         variable = Subscript(variable, Variable(i.name()))
 
@@ -91,20 +87,13 @@ class CudaExpressionBuilder(ExpressionBuilder):
     def buildMultiArraySubscript(self, variable, indices):
         return buildMultiArraySubscript(variable, indices)
 
-    def subscript(self, tree, depth=None):
-        meth = getattr(self, "subscript_"+tree.__class__.__name__)
-        if depth is None:
-            return meth(tree)
-        else:
-            return meth(tree, depth)
-
     def subscript_Argument(self, tree):
         # Build the subscript based on the argument count
         count = tree.count()
         indices = []
         for dimIndices in self._indexStack:
             indices.extend(dimIndices)
-        indices = indices + [self._formBackend.buildBasisIndex(count), self._formBackend.buildGaussIndex()]
+        indices += [self._formBackend.buildBasisIndex(count), self._formBackend.buildGaussIndex()]
         return indices
 
     def subscript_SpatialDerivative(self,tree,dimIndices):
@@ -113,12 +102,12 @@ class CudaExpressionBuilder(ExpressionBuilder):
         operand, _ = tree.operands()
         count = operand.count()
 
-        if isinstance(operand, ufl.argument.Argument):
+        if isinstance(operand, Argument):
             indices = [ ElementIndex()]
             indices.extend(dimIndices)
             indices = indices + [ self._formBackend.buildGaussIndex(),
                                   self._formBackend.buildBasisIndex(count) ]
-        elif isinstance(operand, ufl.coefficient.Coefficient):
+        elif isinstance(operand, Coefficient):
             indices = [ self._formBackend.buildGaussIndex() ]
             indices.extend(dimIndices)
 
@@ -137,22 +126,6 @@ class CudaExpressionBuilder(ExpressionBuilder):
 
         return indices
 
-    def subscript_CoeffQuadrature(self, coeff):
-        # Build the subscript based on the rank
-        indices = [self._formBackend.buildGaussIndex()]
-        rank = coeff.rank()
-
-        if isinstance(coeff, SpatialDerivative):
-            rank = rank + 1 
-
-        if rank > 0:
-            dimIndices = self._indexStack.peek()
-            if len(dimIndices) != rank:
-                raise RuntimeError("Number of indices does not match rank of coefficient. %d vs %d." % (len(dimIndices), rank))
-            indices.extend(dimIndices)
-        
-        return indices
-
 class CudaQuadratureExpressionBuilder(QuadratureExpressionBuilder):
 
     def buildSubscript(self, variable, indices):
@@ -162,16 +135,8 @@ class CudaQuadratureExpressionBuilder(QuadratureExpressionBuilder):
         rank = tree.rank()
         indices = [ ElementIndex() ]
         for r in range(rank):
-            index = self._formBackend.buildDimIndex(r)
-            indices.append(index)
+            indices.append(self._formBackend.buildDimIndex(r))
         indices.append(self._formBackend.buildBasisIndex(0))
-        return indices
-
-    def subscript_argument(self, tree):
-        # The count of the basis function induction variable is always
-        # 0 in the quadrature loops (i.e. i_r_0)
-        indices = [self._formBackend.buildBasisIndex(0), 
-                   self._formBackend.buildGaussIndex()]
         return indices
 
     def subscript_spatial_derivative(self, tree):
